requirements/requirements-include-installer.txt
altgraph==0.16.1
future==0.17.1
lxml==4.2.4
macholib==1.11
pefile==2018.8.8
Pillow==5.4.1
PyInstaller==3.4
PyQt5==5.10.1
pywin32-ctypes==0.2.0
sip==4.19.8
